Add files via upload
Now your friend has followed a lot of different food bloggers, he needs to analyse the habits of these bloggers.
From the list of instagram handles you obtained when you searched ‘food’ in previous project. Open the first 10 handles and find the top 5 which have the highest number of followers
Now Find the number of posts these handles have done in the previous 3 days.
Depict this information using a suitable graph.
Your friend also needs a list of hashtags that he should use in his posts.
Open the 5 handles you obtained in the last question, and scrape the content of the first 10 posts of each handle.
Prepare a list of all words used in all the scraped posts and calculate the frequency of each word.
Create a csv file with two columns : the word and its frequency
Now, find the hashtags that were most popular among these bloggers
Plot a Pie Chart of the top 5 hashtags obtained and the number of times they were used by these bloggers in the scraped posts.
You need to also calculate average followers : likes ratio for the obtained handles.
Followers : Likes ratio is calculated as follows:
Find out the likes of the top 10 posts of the 5 handles obtained earlier.
Calculate the average likes for a handle.
Divide the average likes obtained from the number of followers of the handle to get the average followers:like ratio of each handle.
Create a bar graph to depict the above obtained information.